function/requirements.txt
appdirs==1.4.4
autopep8==1.5.7
azure-functions==1.2.1
certifi==2021.5.30
click==8.0.1
colorama==0.4.4
distlib==0.3.2
filelock==3.0.12
Flask==2.0.1
importlib-metadata==4.6.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
pipenv==2021.5.29
psycopg2==2.9.1
psycopg2-binary==2.9.1
pycodestyle==2.7.0
pymongo==3.10.1
python-dotenv==0.18.0
python-http-client==3.3.2
sendgrid==6.7.1
six==1.16.0
starkbank-ecdsa==1.1.1
toml==0.10.2
typing-extensions==3.10.0.0
virtualenv==20.4.7
virtualenv-clone==0.5.4
Werkzeug==2.0.1
zipp==3.5.0
